Parallelised file creation
* Made a nice solution to creating separate files in parallel. I was
initially using the presence of a file to tell the generator whether or
not to make a new one for the current user (seeing as the same config
will yield the same results). So now I’m making the files sequentially,
but populating them concurrently meaning I can use the same mechanism
whilst creating files in parallel!